gpu virt addr management: does this look sane?
we use drm_mm to mange the virtaul gpu address space. so if we want
to make a gem into the address space of the gpu we use drm_mm to find
a free area and map it via the iommu of the gpu.

I need to say that in theory this should work but my network is broken
today and so I can not test it on real hw.

I think its time to add some cmd stream debugsfs thing.

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>