staging: etnaviv: fix etnaviv gpu debugfs output
"0x08%x" is a confusing format.  It prints values prefixed with 0x08
irrespective of their values, which leads you to think that 0x08fe
really is 0x08fe.  Fix the format string to place the % correctly.

Reverse the 64-bit values - even though the GPU wants 64-bit alignment
in the command stream, and sees the stream as 64-bit values, it's more
humanly understandable to print them as separate 32-bit values in
"lo hi" format - the order that they appear in the command stream.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>