get rid of submit_offset in struct drm_etnaviv_gem_submit_cmd
We do not need this offset as we always submit cmd buffers
starting from the beginning of the buffer. As we need to
patch each submited cmd buffer (jump to next buffer or
jump back to kernel buffer) we need to leave some space
for this in the submited buffer. This adds too much
complexity -> kick it.

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>