add to_bytes(..) helper
This little helper may look very stupid, but it helps
to keep track of places where we want to convert cmd stream
worlds to bytes. I think this makes the code a little bit
easier to understand. As the to_bytes(..) is a static
inline function there is no impact on performance.

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>