add basic infrastructure for so called management events.
They are needed in cases where we want to get notified if a
special point in the cmd stream got processed. It can be
compared to the fence stuff but the management event is used
only by the driver. For instance, we need to wait until the
MMUv2 gots configured via cmd stream before we can enable it.
Or if we want to stop the FE we can use the STOP opcode, which
can trigger an event too. This makes it possible to do active
waiting.

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>