cmd stream validator: register describing a phys address needs to be 0
As we do not want the user space to do bad stuff by writting cmd streams
dealing with physical addresses we force every address register in the
stream to be 0x0.

As the kernel patches the cmd stream later based on reloc information
based on bo's we are safe for this case. I hope I got every register
dealing with addresses.

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>