drm/etnaviv: global perfmon mode
Introduce a global perfmon mode that can be set and get via
DRM_ETNAVIV_SET_PARAM and DRM_ETNAVIV_GET_PARAM.
In global perfmon mode we disable clock gating and enable
debug registers. This is needed if we want to sample perfmon
counter in a global state instead of a per-draw-call view.

We need to keep track of global perfmon mode on a per file basis
to help bogus and/or dying processes to not mess up with our
refcounted perfmon state.

Signed-off-by: Christian Gmeiner <christian.gmeiner@gmail.com>